Don't read JSON from request of non application/json content-type
Some parts of the code already guard against this, but some don't. This
leads to a problem in glance, where accessing `request.json` on a failed
image-upload (with content-type "application/octet-stream") can lead to
the auditmiddleware downloading the whole image body, which - for glance
- automatically ends up in a temporary file, because that's how the
WebOb library used by glance handles big WSGI request bodies.

We now only read `request.json` if the request has the appropriate
content-type.

An alternative would be to prohibit reading `request.json` if the
content-type is "application/octet-stream", but I'm not sure if we have
broken services out there, that support JSON in non-json requests.