Handle requests with no content length
Some requests (like DELETE) have a content_length of None in their
response. As trying to compare None with an integer raises an exception
we need to make sure we have an int to compare to.

Without this patch a DELETE will trigger the following exception:

TypeError: '>' not supported between instances of 'NoneType' and 'int'