Remove duplicate settings in setup.py
Instead of setting description, name etc. in `setup.py` and `setup.cfg`,
we switch to only setting them in `setup.cfg`. While using `setup.py`
for this is still supported, the standard a couple years back was using
`setup.cfg` - and now we moved off to `pyproject.toml`.

The reason for this change is, that automatic Python package building
with `hatch` fails in our CI. Removing these settings makes it work.